Rename passive voice config files to make it clear they are used for the passive voice assessment
