+++ packages/yoastseo/src/languageProcessing/languages/de/config/functionWords.js
@@ -1,6 +1,6 @@
-import filteredPassiveAuxiliariesFactory from "./internal/auxiliaries.js";
+import filteredPassiveAuxiliariesFactory from "./internal/passiveVoiceAuxiliaries.js";
 const filteredPassiveAuxiliaries = filteredPassiveAuxiliariesFactory().filteredAuxiliaries;
-import passiveAuxiliariesInfinitiveFactory from "./internal/auxiliaries.js";
+import passiveAuxiliariesInfinitiveFactory from "./internal/passiveVoiceAuxiliaries.js";
 const passiveAuxiliariesInfinitive = passiveAuxiliariesInfinitiveFactory().infinitiveAuxiliaries;
 import transitionWordsFactory from "./transitionWords.js";
 const transitionWords = transitionWordsFactory().singleWords;
