+++ packages/yoastseo/src/languageProcessing/languages/de/config/internal/passiveVoiceAuxiliaries.js
