+++ packages/yoastseo/src/languageProcessing/languages/de/config/internal/passiveVoiceIrregulars.js
