+++ packages/yoastseo/src/languageProcessing/languages/de/config/internal/passiveVoiceRegex.js
